+++ molviewspec/molviewspec/builder.py
@@ -751,6 +751,25 @@ class Component(_Base, _FocusMixin):
     def representation(
         self,
         *,
+        type: Literal["carbohydrate"],
+        size_factor: float = None,
+        custom: CustomT = None,
+        ref: RefT = None,
+    ) -> Representation:
+        """
+        Add a carbohydrate representation for this component.
+        :param type: the type of this representation ('carbohydrate')
+        :param size_factor: adjust the scale of the visuals (relative to 1.0)
+        :param custom: optional, custom data to attach to this node
+        :param ref: optional, reference that can be used to access this node
+        :return: a builder that handles operations at representation level
+        """
+        ...
+
+    @overload
+    def representation(
+        self,
+        *,
         type: Literal["surface"],
         ignore_hydrogens: bool = None,
         size_factor: float = None,